requirements.txt
click==7.1.1
cycler==0.10.0
Django==2.2.5
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.0
numpy==1.18.1
opencv-contrib-python==3.3.0.10
opencv-python==3.3.0.10
Pillow==7.0.0
pkg-resources==0.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
Werkzeug==1.0.0
WTForms==2.2.1
